Refactor to a new module.
