+++ pybossa/jobs.py
@@ -22,6 +22,7 @@ from flask import current_app, render_template
 from flask.ext.mail import Message
 from pybossa.core import mail, task_repo, importer
 from pybossa.util import with_cache_disabled
+from pybossa.dashboard import get_dashboard_jobs
 
 
 MINUTE = 60
@@ -99,8 +100,10 @@ def get_periodic_jobs(queue):
     engage_jobs = get_inactive_users_jobs() if queue == 'quaterly' else []
     non_contrib_jobs = get_non_contributors_users_jobs() \
         if queue == 'quaterly' else []
+    dashboard_jobs = get_dashboard_jobs() if queue == 'low' else []
     _all = [zip_jobs, jobs, project_jobs, autoimport_jobs,
-            engage_jobs, non_contrib_jobs]
+            engage_jobs, non_contrib_jobs, dashboard_jobs]
+    _all = [dashboard_jobs]
     return (job for sublist in _all for job in sublist if job['queue'] == queue)
 
 
@@ -112,10 +115,6 @@ def get_default_jobs():  # pragma: no cover
                timeout=(10 * MINUTE), queue='low')
     yield dict(name=warm_cache, args=[], kwargs={},
                timeout=(10 * MINUTE), queue='super')
-    yield dict(name=dashboard_active_users_week, args=[], kwargs={},
-              timeout=(10 * MINUTE), queue='low')
-    yield dict(name=dashboard_active_anon_week, args=[], kwargs={},
-              timeout=(10 * MINUTE), queue='low')
 
 
 def get_export_task_jobs(queue):
@@ -413,96 +412,3 @@ def import_tasks(project_id, **form_data):
                      subject=subject, body=body)
     send_mail(mail_dict)
     return msg
-
-
-def dashboard_active_users_week():
-    """Get active users last week."""
-    from sqlalchemy import text
-    from pybossa.core import db
-    # Check first if the materialized view exists
-    sql = text('''SELECT EXISTS (SELECT relname FROM pg_class WHERE
-               relname='dashboard_week_users');''')
-    results = db.slave_session.execute(sql)
-    for row in results:
-        if row.exists:
-            sql = text('''REFRESH MATERIALIZED VIEW
-                       dashboard_week_users''')
-            db.session.execute(sql)
-            return "Materialized view refreshed"
-        else:
-            sql = text('''CREATE MATERIALIZED VIEW dashboard_week_users AS
-                       WITH crafters_per_day AS
-                            (select to_date(task_run.finish_time,
-                                            'YYYY-MM-DD\THH24:MI:SS.US') AS day,
-                                    user_id, COUNT(task_run.user_id) AS day_crafters
-                            FROM task_run
-                            WHERE to_date(task_run.finish_time,
-                                          'YYYY-MM-DD\THH24:MI:SS.US')
-                                >= NOW() - ('1 week'):: INTERVAL
-                            GROUP BY day, task_run.user_id)
-                       SELECT day, COUNT(crafters_per_day.user_id) AS n_users
-                       FROM crafters_per_day GROUP BY day ORDER BY day;''')
-            results = db.slave_session.execute(sql)
-            db.session.commit()
-            return "Materialized view created"
-
-
-def dashboard_active_anon_week():
-    """Get active anon last week."""
-    from sqlalchemy import text
-    from pybossa.core import db
-    # Check first if the materialized view exists
-    sql = text('''SELECT EXISTS (SELECT relname FROM pg_class WHERE
-               relname='dashboard_week_anon');''')
-    results = db.slave_session.execute(sql)
-    for row in results:
-        if row.exists:
-            sql = text('''REFRESH MATERIALIZED VIEW
-                       dashboard_week_anon''')
-            db.session.execute(sql)
-            return "Materialized view refreshed"
-        else:
-            sql = text('''CREATE MATERIALIZED VIEW dashboard_week_anon AS
-                       WITH crafters_per_day AS
-                            (select to_date(task_run.finish_time,
-                                            'YYYY-MM-DD\THH24:MI:SS.US') AS day,
-                                    user_ip, COUNT(task_run.user_ip) AS day_crafters
-                            FROM task_run
-                            WHERE to_date(task_run.finish_time,
-                                          'YYYY-MM-DD\THH24:MI:SS.US')
-                                >= NOW() - ('1 week'):: INTERVAL
-                            GROUP BY day, task_run.user_ip)
-                       SELECT day, COUNT(crafters_per_day.user_ip) AS n_users
-                       FROM crafters_per_day GROUP BY day ORDER BY day;''')
-            results = db.slave_session.execute(sql)
-            db.session.commit()
-            return "Materialized view created"
-
-def dashboard_new_projects_week():
-    """Get new created projects last week."""
-    from sqlalchemy import text
-    from pybossa.core import db
-    # Check first if the materialized view exists
-    sql = text('''SELECT EXISTS (SELECT relname FROM pg_class WHERE
-               relname='dashboard_week_anon');''')
-    results = db.slave_session.execute(sql)
-    for row in results:
-        if row.exists:
-            sql = text('''REFRESH MATERIALIZED VIEW
-                       dashboard_week_project_new''')
-            db.session.execute(sql)
-            return "Materialized view refreshed"
-        else:
-            sql = text('''CREATE MATERIALIZED VIEW dashboard_week_project_new AS
-                       SELECT TO_DATE('YYYY-MM-DD\THH24:MI:SS.US') as day,
-                       project.id, short_name, project.name,
-                       owner_id, "user".name, "user".email_addr
-                       FROM project, "user"
-                       WHERE TO_DATE(project.created,
-                                    'YYYY-MM-DD\THH24:MI:SS.US') >= now() -
-                                    ('1 week')::INTERVAL
-                       AND "user".id=project.owner_id
-                       GROUP BY project.id, "user".name, "user".email_addr;''')
-            results = db.slave_session.execute(sql)
-            db.session.commit()
-            return "Materialized view created"
